+++ structurizr-core/test/unit/com/structurizr/model/RelationshipTests.java
@@ -25,21 +25,21 @@ public class RelationshipTests extends AbstractWorkspaceTestBase {
     @Test
     public void test_getTags_WhenThereAreNoTags() {
         Relationship relationship = softwareSystem1.uses(softwareSystem2, "uses");
-        assertEquals("Relationship,Synchronous", relationship.getTags());
+        assertEquals("Relationship", relationship.getTags());
     }
 
     @Test
     public void test_getTags_ReturnsTheListOfTags_WhenThereAreSomeTags() {
         Relationship relationship = softwareSystem1.uses(softwareSystem2, "uses");
         relationship.addTags("tag1", "tag2", "tag3");
-        assertEquals("Relationship,Synchronous,tag1,tag2,tag3", relationship.getTags());
+        assertEquals("Relationship,tag1,tag2,tag3", relationship.getTags());
     }
 
     @Test
     public void test_setTags_ClearsTheTags_WhenPassedNull() {
         Relationship relationship = softwareSystem1.uses(softwareSystem2, "uses");
         relationship.addTags("Tag 1", "Tag 2");
-        assertEquals("Relationship,Synchronous,Tag 1,Tag 2", relationship.getTags());
+        assertEquals("Relationship,Tag 1,Tag 2", relationship.getTags());
         relationship.setTags(null);
         assertEquals("Relationship", relationship.getTags());
     }
@@ -48,32 +48,36 @@ public class RelationshipTests extends AbstractWorkspaceTestBase {
     public void test_addTags_DoesNotDoAnything_WhenPassedNull() {
         Relationship relationship = softwareSystem1.uses(softwareSystem2, "uses");
         relationship.addTags((String)null);
-        assertEquals("Relationship,Synchronous", relationship.getTags());
+        assertEquals("Relationship", relationship.getTags());
 
         relationship.addTags(null, null, null);
-        assertEquals("Relationship,Synchronous", relationship.getTags());
+        assertEquals("Relationship", relationship.getTags());
     }
 
     @Test
     public void test_addTags_AddsTags_WhenPassedSomeTags() {
         Relationship relationship = softwareSystem1.uses(softwareSystem2, "uses");
         relationship.addTags(null, "tag1", null, "tag2");
-        assertEquals("Relationship,Synchronous,tag1,tag2", relationship.getTags());
+        assertEquals("Relationship,tag1,tag2", relationship.getTags());
     }
 
     @Test
-    public void test_getInteractionStyle_ReturnsSynchronous_WhenNotExplicitlySet() {
+    public void test_getInteractionStyle_ReturnsNull_WhenNotExplicitlySet() {
         Relationship relationship = softwareSystem1.uses(softwareSystem2, "uses");
-        assertEquals(InteractionStyle.Synchronous, relationship.getInteractionStyle());
+        assertNull(relationship.getInteractionStyle());
     }
 
     @Test
     public void test_getTags_IncludesTheInteractionStyleWhenSpecified() {
         Relationship relationship = softwareSystem1.uses(softwareSystem2, "Uses 1", "Technology");
+        assertFalse(relationship.getTags().contains(Tags.SYNCHRONOUS));
+        assertFalse(relationship.getTags().contains(Tags.ASYNCHRONOUS));
+
+        relationship.setInteractionStyle(InteractionStyle.Synchronous);
         assertTrue(relationship.getTags().contains(Tags.SYNCHRONOUS));
         assertFalse(relationship.getTags().contains(Tags.ASYNCHRONOUS));
 
-        relationship = softwareSystem1.uses(softwareSystem2, "Uses 2", "Technology", InteractionStyle.Asynchronous);
+        relationship.setInteractionStyle(InteractionStyle.Asynchronous);
         assertFalse(relationship.getTags().contains(Tags.SYNCHRONOUS));
         assertTrue(relationship.getTags().contains(Tags.ASYNCHRONOUS));
     }